requirements.txt
boto3
pytest
PyYAML